+++ packages/base/src/Device/index.ts
@@ -1,18 +1,10 @@
 import '@ui5/webcomponents-react-base/types/UI5Device.d.ts';
-import { getBrowser, getOS, getSystem, supportTouch } from '@ui5/webcomponents-base/dist/Device';
-import { deprecationNotice } from '@ui5/webcomponents-react-base/lib/Utils';
-import { EventRegistry } from './EventRegistry';
-import { MediaLegacy } from './Media';
-import {
-  supportInputPlaceholder,
-  supportMatchMedia,
-  supportMatchMediaListener,
-  supportOrientation,
-  supportPointerEvents,
-  supportRetina,
-  supportWebSocket
-} from './Support';
+import { getOS, getSystem, supportTouch } from '@ui5/webcomponents-base/dist/Device';
+import EventProvider from '@ui5/webcomponents-base/dist/EventProvider';
+import { supportOrientation } from './Support';
 import * as Utils from './utils';
+
+const eventProvider = new EventProvider();
 
 let iResizeTimeout;
 let bOrientationChange = false;
@@ -34,12 +26,12 @@ interface IWindowSize {
   width: number;
 }
 
-const windowSize: IWindowSize = {
+const internalWindowSize: IWindowSize = {
   height: 0,
   width: 0
 };
 
-const orientation: IOrientation = {
+const internalOrientation: IOrientation = {
   landscape: false,
   portrait: false
 };
@@ -47,13 +39,13 @@ const orientation: IOrientation = {
 // PRIVATE API
 
 const setResizeInfo = () => {
-  windowSize.width = Utils.getWindowSize()[0];
-  windowSize.height = Utils.getWindowSize()[1];
+  internalWindowSize.width = Utils.getWindowSize()[0];
+  internalWindowSize.height = Utils.getWindowSize()[1];
 };
 
 const setOrientationInfo = () => {
-  orientation.landscape = Utils.isLandscape(true, orientation, bKeyboardOpen);
-  orientation.portrait = !orientation.landscape;
+  internalOrientation.landscape = Utils.isLandscape(true, internalOrientation, bKeyboardOpen);
+  internalOrientation.portrait = !internalOrientation.landscape;
 };
 
 const clearFlags = () => {
@@ -81,7 +73,10 @@ const initEventListeners = () => {
 // orientation change
 const handleOrientationChange = () => {
   setOrientationInfo();
-  EventRegistry.fireEvent('orientation', { landscape: orientation.landscape, portrait: orientation.portrait });
+  eventProvider.fireEvent('orientation', {
+    landscape: internalOrientation.landscape,
+    portrait: internalOrientation.portrait
+  });
 };
 
 const handleMobileTimeout = () => {
@@ -154,9 +149,9 @@ const handleMobileOrientationResizeChange = (evt) => {
 // RESIZE ONLY WITHOUT ORIENTATION CHANGE
 const handleResizeChange = () => {
   setResizeInfo();
-  EventRegistry.fireEvent('resize', {
-    height: windowSize.height,
-    width: windowSize.width
+  eventProvider.fireEvent('resize', {
+    height: internalWindowSize.height,
+    width: internalWindowSize.width
   });
 };
 
@@ -166,8 +161,8 @@ const handleResizeTimeout = () => {
 };
 
 const handleResizeEvent = () => {
-  const wasL = orientation.landscape;
-  const isL = Utils.isLandscape(false, orientation, bKeyboardOpen);
+  const wasL = internalOrientation.landscape;
+  const isL = Utils.isLandscape(false, internalOrientation, bKeyboardOpen);
   if (wasL !== isL) {
     handleOrientationChange();
   }
@@ -179,245 +174,39 @@ const handleResizeEvent = () => {
   }
 };
 
-class DeviceBuilder {
-  public get os() {
-    deprecationNotice(
-      'Device.os',
-      `Accessing 'Device.os' is deprecated and will be removed in the next major version. 
-    Please use 'getOS()' ('import { getOS } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-    );
-    return getOS();
-  }
-
-  public get browser() {
-    deprecationNotice(
-      'Device.browser',
-      `Accessing 'Device.browser' is deprecated and will be removed in the next major version. 
-    Please use 'getBrowser()' ('import { getBrowser } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-    );
-    return getBrowser();
-  }
-
-  public get system() {
-    deprecationNotice(
-      'Device.system',
-      `Accessing 'Device.system' is deprecated and will be removed in the next major version. 
-    Please use 'getSystem()' ('import { getSystem } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-    );
-    return getSystem();
-  }
-
-  public support = {
-    get touch() {
-      deprecationNotice(
-        'Device.support.touch',
-        `Accessing 'Device.support.touch' is deprecated and will be removed in the next major version. 
-    Please use 'supportTouch()' ('import { supportTouch } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return supportTouch();
-    },
-    get pointer() {
-      deprecationNotice(
-        'Device.support.pointer',
-        `Accessing 'Device.support.pointer' is deprecated and will be removed in the next major version. 
-    Please use 'supportPointerEvents()' ('import { supportPointerEvents } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return supportPointerEvents();
-    },
-    get matchmedia() {
-      deprecationNotice(
-        'Device.support.matchmedia',
-        `Accessing 'Device.support.matchmedia' is deprecated and will be removed in the next major version. 
-    Please use 'supportMatchMedia()' ('import { supportMatchMedia } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return supportMatchMedia();
-    },
-    get matchmedialistener() {
-      deprecationNotice(
-        'Device.support.matchmedialistener',
-        `Accessing 'Device.support.matchmedialistener' is deprecated and will be removed in the next major version. 
-    Please use 'supportMatchMediaListener()' ('import { supportMatchMediaListener } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return supportMatchMediaListener();
-    },
-    get orientation() {
-      deprecationNotice(
-        'Device.support.orientation',
-        `Accessing 'Device.support.orientation' is deprecated and will be removed in the next major version. 
-    Please use 'supportOrientation()' ('import { supportOrientation } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return supportOrientation();
-    },
-    get retina() {
-      deprecationNotice(
-        'Device.support.retina',
-        `Accessing 'Device.support.retina' is deprecated and will be removed in the next major version. 
-    Please use 'supportRetina()' ('import { supportRetina } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return supportRetina();
-    },
-    get websocket() {
-      deprecationNotice(
-        'Device.support.websocket',
-        `Accessing 'Device.support.websocket' is deprecated and will be removed in the next major version. 
-    Please use 'supportWebSocket()' ('import { supportWebSocket } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return supportWebSocket();
-    },
-    input: {
-      get placeholder() {
-        deprecationNotice(
-          'Device.support.input.placeholder',
-          `Accessing 'Device.support.input.placeholder' is deprecated and will be removed in the next major version. 
-    Please use 'supportInputPlaceholder()' ('import { supportInputPlaceholder } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-        );
-        return supportInputPlaceholder();
-      }
-    }
-  };
-
-  public get media() {
-    return new MediaLegacy();
-  }
-
-  public orientation = {
-    get landscape() {
-      deprecationNotice(
-        'Device.orientation.landscape',
-        `Accessing 'Device.orientation.landscape' is deprecated and will be removed in the next major version. 
-    Please use 'getOrientation().landscape' ('import { getOrientation } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return orientation.landscape;
-    },
-    get portrait() {
-      deprecationNotice(
-        'Device.orientation.portrait',
-        `Accessing 'Device.orientation.portrait' is deprecated and will be removed in the next major version. 
-    Please use 'getOrientation().portrait' ('import { getOrientation } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return orientation.portrait;
-    },
-    get attachHandler() {
-      deprecationNotice(
-        'Device.orientation.attachHandler',
-        `Accessing 'Device.orientation.attachHandler' is deprecated and will be removed in the next major version. 
-    Please use 'attachOrientationChangeHandler' ('import { attachOrientationChangeHandler } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return attachOrientationChangeHandler;
-    },
-    get detachHandler() {
-      deprecationNotice(
-        'Device.orientation.detachHandler',
-        `Accessing 'Device.orientation.detachHandler' is deprecated and will be removed in the next major version. 
-    Please use 'detachOrientationChangeHandler' ('import { detachOrientationChangeHandler } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return detachOrientationChangeHandler;
-    }
-  };
-  public resize = {
-    get width() {
-      deprecationNotice(
-        'Device.resize.width',
-        `Accessing 'Device.resize.width' is deprecated and will be removed in the next major version. 
-    Please use 'getWindowSize().width' ('import { getWindowSize } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return windowSize.width;
-    },
-    get portrait() {
-      deprecationNotice(
-        'Device.resize.height',
-        `Accessing 'Device.resize.height' is deprecated and will be removed in the next major version. 
-    Please use 'getWindowSize().height' ('import { getWindowSize } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return windowSize.height;
-    },
-    get setResizeInfo() {
-      deprecationNotice(
-        'Device.resize.setResizeInfo',
-        `Accessing 'Device.resize.setResizeInfo' is deprecated and will be removed in the next major version. 
-        There will be no replacement as this was a private API.`
-      );
-      return setResizeInfo;
-    },
-    get attachHandler() {
-      deprecationNotice(
-        'Device.resize.attachHandler',
-        `Accessing 'Device.resize.attachHandler' is deprecated and will be removed in the next major version. 
-    Please use 'attachResizeHandler' ('import { attachResizeHandler } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return attachResizeHandler;
-    },
-    get detachHandler() {
-      deprecationNotice(
-        'Device.resize.detachHandler',
-        `Accessing 'Device.resize.detachHandler' is deprecated and will be removed in the next major version. 
-    Please use 'detachResizeHandler' ('import { detachResizeHandler } from '@ui5/webcomponents-react-base/lib/Device') instead.`
-      );
-      return detachResizeHandler;
-    }
-  };
-
-  constructor() {
-    if (!eventListenersInitialized) {
-      initEventListeners();
-    }
-  }
-
-  // LEGACY API
-  private setOrientationInfo = setOrientationInfo;
-  private isLandscape = (bFromOrientationChange) =>
-    Utils.isLandscape(bFromOrientationChange, orientation, bKeyboardOpen);
-
-  // orientation
-  private handleOrientationChange = handleOrientationChange;
-  private handleMobileTimeout = handleMobileTimeout;
-  private handleMobileOrientationResizeChange = handleMobileOrientationResizeChange;
-
-  // resize
-  private handleResizeEvent = handleResizeEvent;
-  private handleResizeChange = handleResizeChange;
-  private handleResizeTimeout = handleResizeTimeout;
-}
-
-export const Device = new DeviceBuilder();
-
 // re-export everything from the web components device
 export * from '@ui5/webcomponents-base/dist/Device';
 // export all support methods
 export * from './Support';
+// export all media methods
+export * from './Media';
 // resize events
 export const getWindowSize = () => {
-  return windowSize;
+  return internalWindowSize;
 };
-export const attachResizeHandler = (fnFunction: (windowSize: IWindowSize) => void, oListener?: unknown): void => {
+export const attachResizeHandler = (fnFunction: (windowSize: IWindowSize) => void): void => {
   if (!eventListenersInitialized) {
     initEventListeners();
   }
-  EventRegistry.attachEvent('resize', fnFunction, oListener);
+  eventProvider.attachEvent('resize', fnFunction);
 };
 
-export const detachResizeHandler = (fnFunction: (windowSize: IWindowSize) => void, oListener?: unknown) => {
-  EventRegistry.detachEvent('resize', fnFunction, oListener);
+export const detachResizeHandler = (fnFunction: (windowSize: IWindowSize) => void) => {
+  eventProvider.detachEvent('resize', fnFunction);
 };
 
 // orientation change events
 export const getOrientation = () => {
-  return orientation;
+  return internalOrientation;
 };
 
-export const attachOrientationChangeHandler = (
-  fnFunction: (orientation: IOrientation) => void,
-  oListener?: unknown
-): void => {
+export const attachOrientationChangeHandler = (fnFunction: (orientation: IOrientation) => void): void => {
   if (!eventListenersInitialized) {
     initEventListeners();
   }
-  EventRegistry.attachEvent('orientation', fnFunction, oListener);
+  eventProvider.attachEvent('orientation', fnFunction);
 };
 
-export const detachOrientationChangeHandler = (
-  fnFunction: (orientation: IOrientation) => void,
-  oListener?: unknown
-) => {
-  EventRegistry.detachEvent('orientation', fnFunction, oListener);
+export const detachOrientationChangeHandler = (fnFunction: (orientation: IOrientation) => void) => {
+  eventProvider.detachEvent('orientation', fnFunction);
 };