+++ packages/main/src/lib/AvatarGroupType.ts
@@ -1,0 +1,1 @@
+export { AvatarGroupType } from '../enums/AvatarGroupType';