+++ packages/main/src/webComponents/NotificationListGroupItem/index.tsx
@@ -39,7 +39,7 @@ export interface NotificationListGroupItemPropTypes extends WithWebComponentProp
   /**
    * Defines the actions, displayed in the top-right area.
    *
-   * **Note:** use the `NotificationOverflowAction` component.
+   * **Note:** use the `NotificationAction` component.
    */
   actions?: ReactNode;
   /**
@@ -59,14 +59,12 @@ export interface NotificationListGroupItemPropTypes extends WithWebComponentProp
  * *   `Toggle` button to expand and collapse the group
  * *   `Priority` icon to display the priority of the group
  * *   `Heading` to entitle the group
- * *   Custom actions - with the use of `NotificationOverflowAction`
+ * *   Custom actions - with the use of `NotificationAction`
  * *   Items of the group
  *
  * <a href="https://sap.github.io/ui5-webcomponents/playground/components/NotificationListGroupItem" target="_blank">UI5 Web Components Playground</a>
  */
-const NotificationListGroupItem: FC<NotificationListGroupItemPropTypes> = withWebComponent<
-  NotificationListGroupItemPropTypes
->(
+const NotificationListGroupItem: FC<NotificationListGroupItemPropTypes> = withWebComponent<NotificationListGroupItemPropTypes>(
   'ui5-li-notification-group',
   ['heading', 'priority'],
   ['collapsed', 'showCounter', 'busy', 'showClose', 'selected'],